+++ src/middleware/asyncHandler.js
@@ -3,6 +3,8 @@
  * @param {Function} fn - Asynchronous route handler function.
  * @returns n/a. This middleware allows Express to catch asynchronous errors that it cannot otherwise do.
  */
-export const asyncHandler = (fn) => (req, res, next) => {
+const asyncHandler = (fn) => (req, res, next) => {
     Promise.resolve(fn(req, res, next)).catch(next)
 }
+
+export default asyncHandler
