+++ src/middleware/sessionAuth.js
@@ -3,10 +3,14 @@
  * Applied to any routes that require access to an authenticated session.
  * Verifies that there is an active session tied to the session ID presented by the client's cookie.
  */
-export const sessionAuth = (req, res, next) => {
+const sessionAuth = (req, res, next) => {
     if (!req.session.user) {
         return res.status(401).end('Unauthorised')
     }
 
+    console.log('Active session found.')
+
     next()
 }
+
+export default sessionAuth
