+++ src/middleware/tokenExpiry.js
@@ -1,12 +1,12 @@
 import { spotifyAPI } from '../index.js'
-import { calculateExpiryUTC } from '../utils.js'
+import calculateExpiryUTC from '../common/utils/calculateExpiryUTC.js'
 
 /**
  * Token expiry middleware
  * Applied to any routes that require the use of an access token.
  * Verifies that the sessions' currently stored access token has not expired. Refreshes the token if yes.
  */
-export const tokenExpiry = async (req, res, next) => {
+const tokenExpiry = async (req, res, next) => {
     const { user } = req.session
     const current_utc = new Date(new Date().toUTCString())
     const expiry_utc = new Date(user.expiry_utc)
@@ -14,6 +14,8 @@ export const tokenExpiry = async (req, res, next) => {
     if (current_utc < expiry_utc) {
         return next()
     }
+
+    console.log('Token expired, refreshing...')
 
     const response = await spotifyAPI.refreshAccessToken(user.refresh_token)
     const { access_token, expires_in } = response.data
@@ -23,3 +25,5 @@ export const tokenExpiry = async (req, res, next) => {
 
     next()
 }
+
+export default tokenExpiry
